remove ApiRouter
