+++ apps/OrderProcessor/src/main/java/com/tradingplatform/orderprocessor/orders/OrderType.java
@@ -1,0 +1,6 @@
+package com.tradingplatform.orderprocessor.orders;
+
+public enum OrderType {
+  BUY,
+  SELL,
+}
