+++ apps/OrderProcessor/src/main/java/com/tradingplatform/orderprocessor/orders/Ticker.java
@@ -1,0 +1,10 @@
+package com.tradingplatform.orderprocessor.orders;
+
+public enum Ticker {
+  AAPL,
+  MSFT,
+  GOOGL,
+  AMZN,
+  TSLA,
+  META,
+}
