--- apps/OrderProcessor/src/main/java/orderProcessor/OrderType.java
@@ -1,6 +1,0 @@
-package orderProcessor;
-
-public enum OrderType {
-  BUY,
-  SELL,
-}
