--- apps/OrderProcessor/src/main/java/orderProcessor/Ticker.java
@@ -1,10 +1,0 @@
-package orderProcessor;
-
-public enum Ticker {
-  AAPL,
-  MSFT,
-  GOOGL,
-  AMZN,
-  TSLA,
-  META,
-}
