--- apps/OrderProcessor/src/test/java/orderProcessor/OrderProcessorTest.java
@@ -1,162 +1,0 @@
-package orderProcessor;
-
-import static org.junit.jupiter.api.Assertions.*;
-
-import java.util.PriorityQueue;
-import org.junit.jupiter.api.*;
-
-class OrderProcessorTest {
-
-  private OrderProcessor orderProcessor;
-
-  @BeforeEach
-  void setUp() {
-    orderProcessor = OrderProcessor.getInstance();
-  }
-
-  @AfterEach
-  void reset() {
-    orderProcessor.resetInstance();
-  }
-
-  @Test
-  void testOrderCancellation() {
-    Order order = new Order(OrderType.SELL, "dummyID", Ticker.AAPL, 110.0, 50);
-    orderProcessor.processOrder(order);
-    Boolean result = orderProcessor.cancelOrder(order.getId().toString(), order.getTicker().name(), order.getType().name());
-    assertTrue(result);
-  }
-
-  @Test
-  void addBuyOrder() {
-    Order order = new Order(OrderType.BUY, "dummyID", Ticker.AAPL, 100.0, 2000);
-    orderProcessor.processOrder(order);
-
-    TradeBook book = orderProcessor.getTradeBook(Ticker.AAPL);
-
-    PriorityQueue<Order> buyBook = book.getBuyOrders();
-    assertTrue(buyBook.contains(order));
-  }
-
-  @Test
-  void addSellOrder() {
-    Order order = new Order(
-      OrderType.SELL,
-      "dummyID",
-      Ticker.AAPL,
-      100.0,
-      2000
-    );
-    orderProcessor.processOrder(order);
-
-    TradeBook book = orderProcessor.getTradeBook(Ticker.AAPL);
-
-    PriorityQueue<Order> sellBook = book.getSellOrders();
-    assertTrue(sellBook.contains(order));
-  }
-
-  @Test
-  void inTimeOrder() {
-    Order oldBuyOrder = new Order(
-      OrderType.BUY,
-      "dummyID",
-      Ticker.AAPL,
-      100.0,
-      2
-    );
-    Order newBuyOrder = new Order(
-      OrderType.BUY,
-      "dummyID",
-      Ticker.AAPL,
-      100.0,
-      2000
-    );
-
-    orderProcessor.processOrder(newBuyOrder);
-    orderProcessor.processOrder(oldBuyOrder);
-
-    TradeBook book = orderProcessor.getTradeBook(Ticker.AAPL);
-    PriorityQueue<Order> buyQueue = book.getBuyOrders();
-
-    assertEquals(
-      oldBuyOrder,
-      buyQueue.poll(),
-      "Older order should be processed first"
-    );
-    assertEquals(
-      newBuyOrder,
-      buyQueue.poll(),
-      "Newer order should be processed second"
-    );
-  }
-
-  @Test
-  void sellByPriceDesc() {
-    Order cheapSellOrder = new Order(
-      OrderType.SELL,
-      "dummyID",
-      Ticker.AAPL,
-      95.0,
-      2000
-    );
-    Order expensiveSellOrder = new Order(
-      OrderType.SELL,
-      "dummyID",
-      Ticker.AAPL,
-      105.0,
-      2000
-    );
-
-    orderProcessor.processOrder(cheapSellOrder);
-    orderProcessor.processOrder(expensiveSellOrder);
-
-    TradeBook book = orderProcessor.getTradeBook(Ticker.AAPL);
-    PriorityQueue<Order> sellBook = book.getSellOrders();
-
-    assertEquals(
-      cheapSellOrder,
-      sellBook.poll(),
-      "Cheap sell order should be processed first"
-    );
-    assertEquals(
-      expensiveSellOrder,
-      sellBook.poll(),
-      "Expensive sell order should be processed second"
-    );
-  }
-
-  @Test
-  void buyByPriceAsc() {
-    Order cheapBuyOrder = new Order(
-      OrderType.BUY,
-      "dummyID",
-      Ticker.AAPL,
-      100.0,
-      2000
-    );
-    Order expensiveBuyOrder = new Order(
-      OrderType.BUY,
-      "dummyID",
-      Ticker.AAPL,
-      105.0,
-      2000
-    );
-
-    orderProcessor.processOrder(cheapBuyOrder);
-    orderProcessor.processOrder(expensiveBuyOrder);
-
-    TradeBook book = orderProcessor.getTradeBook(Ticker.AAPL);
-    PriorityQueue<Order> buyQueue = book.getBuyOrders();
-
-    assertEquals(
-      expensiveBuyOrder,
-      buyQueue.poll(),
-      "Expensive sell order should be processed first"
-    );
-    assertEquals(
-      cheapBuyOrder,
-      buyQueue.poll(),
-      "Cheap sell order should be processed second"
-    );
-  }
-}
